Rename the files the appropriate class name
