+++ source/Makefile
@@ -28,7 +28,8 @@ APP_MIME_SIG =
 #	means this Makefile will not work correctly if two source files with the
 #	same name (source.c or source.cpp) are included from different directories.
 #	Also note that spaces in folder names do not work well with this Makefile.
-SRCS =  main.cpp fenetre.cpp mainview.cpp option.cpp optionview.cpp mousew.cpp mouseview.cpp
+SRCS =  main.cpp MainWindow.cpp MainView.cpp OptionWindow.cpp OptionView.cpp \
+		MouseWindow.cpp MouseView.cpp
 
 #	Specify the resource definition files to use. Full or relative paths can be
 #	used.
