+++ source/OptionView.cpp
@@ -1,4 +1,4 @@
-#include "optionview.h"
+#include "OptionView.h"
 #include <String.h>
 #include <Application.h>
 #include <LayoutBuilder.h>
